Fix for all NANs in contour

+++ lib/matplotlib/contour.py
@@ -1447,12 +1447,12 @@ class QuadContourSet(ContourSet):
         else:
             raise _api.nargs_error(fn, takes="from 1 to 4", given=nargs)
         z = ma.masked_invalid(z, copy=False)
-        self.zmax = float(z.max())
-        self.zmin = float(z.min())
+        self.zmax = z.max().astype(float)
+        self.zmin = z.min().astype(float)
         if self.logscale and self.zmin <= 0:
             z = ma.masked_where(z <= 0, z)
             _api.warn_external('Log scale: values of z <= 0 have been masked')
-            self.zmin = float(z.min())
+            self.zmin = z.min().astype(float)
         self._process_contour_level_args(args, z.dtype)
         return (x, y, z)
